Use 128bit instead of 64bit amounts (#1081)
* Update Espresso deps for 128-bit amounts

* Use 128bit amounts: contract

The TestCAPE contract is now 614 bytes too large.

[ci skip]

* Patch geth to allow contracts up to 32kB in size

move comment

* Use 128bit amounts in wallet

* [WIP] faucet

* Update faucet to use 128bit amounts

* Update reef to 0.2.2

* More consistency

+++ contracts/contracts/AssetRegistry.sol
@@ -31,7 +31,7 @@ contract AssetRegistry {
         EdOnBN254.EdOnBN254Point credPk;
         EdOnBN254.EdOnBN254Point freezerPk;
         uint256 revealMap;
-        uint64 revealThreshold;
+        uint128 revealThreshold;
     }
 
     /// @notice Return the CAP-native asset definition.
